Split shell loader responsibilities
